+++ packages/pluggableWidgets/dropdown-sort-web/src/DropdownSort.editorConfig.ts
@@ -4,7 +4,8 @@ import {
     ContainerProps,
     ImageProps,
     StructurePreviewProps,
-    TextProps
+    text,
+    structurePreviewPalette
 } from "@mendix/pluggable-widgets-commons";
 import { DropdownSortPreviewProps } from "../typings/DropdownSortProps";
 
@@ -12,6 +13,7 @@ import AscIcon from "./assets/asc.svg";
 import AscIconDark from "./assets/asc-dark.svg";
 
 export const getPreview = (values: DropdownSortPreviewProps, isDarkMode: boolean): StructurePreviewProps => {
+    const palette = structurePreviewPalette[isDarkMode ? "dark" : "light"];
     return {
         type: "RowLayout",
         borders: true,
@@ -22,24 +24,16 @@ export const getPreview = (values: DropdownSortPreviewProps, isDarkMode: boolean
             {
                 type: "RowLayout",
                 columnSize: "grow",
-                backgroundColor: isDarkMode ? "#313131" : "#FFFFFF",
+                backgroundColor: palette.background.container,
                 children: [
                     {
                         type: "Container",
                         padding: 8,
                         children: [
-                            {
-                                type: "Text",
-                                fontColor: values.emptyOptionCaption
-                                    ? isDarkMode
-                                        ? "#A4A4A4"
-                                        : "#BBBBBB"
-                                    : isDarkMode
-                                    ? "#313131"
-                                    : "#FFF",
-                                italic: true,
-                                content: values.emptyOptionCaption ? values.emptyOptionCaption : "Sample"
-                            } as TextProps
+                            text({
+                                fontColor: palette.text.secondary,
+                                italic: true
+                            })(values.emptyOptionCaption || " ")
                         ],
                         grow: 1
                     } as ContainerProps,
@@ -65,7 +59,7 @@ export const getPreview = (values: DropdownSortPreviewProps, isDarkMode: boolean
             {
                 type: "Container",
                 grow: 0,
-                backgroundColor: isDarkMode ? "#313131" : "#FFFFFF",
+                backgroundColor: palette.background.container,
                 children: [
                     {
                         type: "Container",