+++ packages/pluggableWidgets/tree-node-web/src/TreeNode.editorConfig.ts
@@ -3,7 +3,8 @@ import {
     DropZoneProps,
     RowLayoutProps,
     StructurePreviewProps,
-    TextProps
+    TextProps,
+    structurePreviewPalette
 } from "@mendix/pluggable-widgets-commons";
 import { hidePropertiesIn, hidePropertyIn, Properties, transformGroupsIntoTabs } from "@mendix/pluggable-widgets-tools";
 
@@ -50,10 +51,12 @@ export function getProperties(
 }
 
 export function getPreview(values: TreeNodePreviewProps, isDarkMode: boolean): StructurePreviewProps | null {
+    const palette = structurePreviewPalette[isDarkMode ? "dark" : "light"];
+
     const titleHeader: RowLayoutProps = {
         type: "RowLayout",
         columnSize: "fixed",
-        backgroundColor: isDarkMode ? "#3B5C8F" : "#DAEFFB",
+        backgroundColor: palette.background.topbarData,
         borders: true,
         borderWidth: 1,
         children: [
@@ -64,7 +67,7 @@ export function getPreview(values: TreeNodePreviewProps, isDarkMode: boolean): S
                     {
                         type: "Text",
                         content: "Tree node",
-                        fontColor: isDarkMode ? "#6DB1FE" : "#2074C8"
+                        fontColor: palette.text.primary
                     }
                 ]
             }
@@ -72,7 +75,7 @@ export function getPreview(values: TreeNodePreviewProps, isDarkMode: boolean): S
     };
     const treeNodeHeader: RowLayoutProps = {
         type: "RowLayout",
-        backgroundColor: isDarkMode ? "#3E3E3E" : "#F5F5F5",
+        backgroundColor: palette.background.topbarStandard,
         borders: true,
         borderWidth: 1,
         children: [
@@ -95,7 +98,7 @@ export function getPreview(values: TreeNodePreviewProps, isDarkMode: boolean): S
                         : ({
                               type: "RowLayout",
                               borders: true,
-                              backgroundColor: isDarkMode ? "#4F4F4F" : "#F8F8F8",
+                              grow: 100,
                               children: [
                                   {
                                       type: "DropZone",