+++ packages/pluggableWidgets/datagrid-web/CHANGELOG.md
@@ -14,6 +14,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 -   We updated the light and dark icons and tiles for the widget.
 
+-   We changed colors in the structure mode preview for dark and light modes.
+
 ### Fixed
 
 -   We fix virtual scrolling issue