+++ packages/pluggableWidgets/fieldset-web/CHANGELOG.md
@@ -6,8 +6,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Changed
+
 -   We updated page explorer's caption to display legend.
+
 -   We updated the light and dark icons and tiles for the widget.
+
+-   We changed colors in the structure mode preview for dark and light modes.
 
 ## [3.1.1] - 2022-10-15